pyproject.toml
[build-system]
requires = ["setuptools>=62.3"]
build-backend = "setuptools.build_meta"

[project]
name = "aioautomower"
# The version is set by GH action on release
version = "0.0.0"
license     = {text = "Apache-2.0"}
description = "MPython module to talk to Husqvarna Automower."
readme = "README.md"
requires-python = ">=3.10"
authors     = [
    {name = "Thomas Protzner", email = "thomas.protzner@gmail.com"}
]
classifiers = [
  "Environment :: Console",
  "Programming Language :: Python :: 3",
]
dependencies = [
  "aiohttp",
  "pydantic",
  "PyJWT"
]


[tool.setuptools]
platforms = ["any"]
zip-safe  = false
packages = ["aioautomower"]
include-package-data = true

[tool.setuptools.package-data]
aioautomower = ["py.typed"]

[tool.ruff]
fix = true
show-fixes = true
exclude = [
    "session.py",
]

# enable later: "C90", "PTH", "TCH", "RET", "ANN"
select = ["E", "F", "W", "I", "N", "D", "UP", "PL", "Q", "SIM", "TID", "ARG"]
ignore = ["PLR2004", "N818"]
unfixable = ["F841"]
line-length = 100
target-version = "py311"

[tool.ruff.flake8-annotations]
allow-star-arg-any = true
suppress-dummy-args = true

[tool.ruff.flake8-builtins]
builtins-ignorelist = ["id"]

[tool.ruff.pydocstyle]
# Use Google-style docstrings.
convention = "pep257"

[tool.ruff.pylint]

max-branches=25
max-returns=15
max-args=10
max-statements=50